#1 - refactor user to users namespace and improve folder structure
