+++ template/backend/src/Ambev.DeveloperEvaluation.Application/Users/CreateUser/CreateUserHandler.cs
@@ -2,8 +2,8 @@
 using MediatR;
 using FluentValidation;
 using Ambev.DeveloperEvaluation.Domain.Repositories;
-using Ambev.DeveloperEvaluation.Domain.Entities;
 using Ambev.DeveloperEvaluation.Common.Security;
+using Ambev.DeveloperEvaluation.Domain.Entities.User;
 
 namespace Ambev.DeveloperEvaluation.Application.Users.CreateUser;
