+++ template/backend/src/Ambev.DeveloperEvaluation.Domain/Events/UserRegisteredEvent.cs
@@ -1,4 +1,4 @@
-using Ambev.DeveloperEvaluation.Domain.Entities;
+using Ambev.DeveloperEvaluation.Domain.Entities.User;
 
 namespace Ambev.DeveloperEvaluation.Domain.Events
 {
