+++ template/backend/src/Ambev.DeveloperEvaluation.Domain/Repositories/IUserRepository.cs
@@ -1,4 +1,4 @@
-using Ambev.DeveloperEvaluation.Domain.Entities;
+using Ambev.DeveloperEvaluation.Domain.Entities.User;
 
 namespace Ambev.DeveloperEvaluation.Domain.Repositories;
