+++ template/backend/src/Ambev.DeveloperEvaluation.Domain/Validation/UserValidator.cs
@@ -1,4 +1,4 @@
-﻿using Ambev.DeveloperEvaluation.Domain.Entities;
+﻿using Ambev.DeveloperEvaluation.Domain.Entities.Users;
 using Ambev.DeveloperEvaluation.Domain.Enums;
 using FluentValidation;
