+++ template/backend/tests/Ambev.DeveloperEvaluation.Unit/Domain/Specifications/TestData/ActiveUserSpecificationTestData.cs
@@ -1,4 +1,4 @@
-using Ambev.DeveloperEvaluation.Domain.Entities;
+using Ambev.DeveloperEvaluation.Domain.Entities.User;
 using Ambev.DeveloperEvaluation.Domain.Enums;
 using Bogus;
